BE-feat: add authorization for category controller

+++ backend/app/Http/Controllers/V1/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\V1;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\Gate;
 
 use App\Models\Category;
 
@@ -13,6 +14,10 @@ class CategoryController extends Controller
 
     public function create(Request $req) {
         try {
+            if (!Gate::check('create', Category::class)) {
+                return $this->sendError(message: 'You have no permission for this action.!', statusCode: 403);
+            }
+
             $fields = $req->only(['name', 'image_url']);
             $validated = Validator::make($fields, [
                 'name' => 'required|string|unique:categories,name|min:5',
@@ -22,6 +27,8 @@ class CategoryController extends Controller
             if ($validated->fails()) {
                 return $this->sendError(message: $validated->messages(), statusCode: 400);
             }
+
+            $fields['user_id'] = auth()->user()->id;
 
             $createdNew = new Category();
             $createdNew->fill($fields);
@@ -35,7 +42,17 @@ class CategoryController extends Controller
 
     public function list(Request $req) {
         try {
-            return $this->sendResponse(message: 'Retrieve all Category success', data: Category::all());
+            $withOwner = $req->query('owner') === 'true';
+
+            if ($withOwner) {
+                $data = auth()->user()->categories;
+                $message = 'Retrieve all Category created by current user success';
+            } else {
+                $data = Category::all();
+                $message = 'Retrieve all Category success';
+            }
+
+            return $this->sendResponse(message: $message, data: $data);
         } catch (\Throwable $th) {
             return $this->sendError(message: $th->getMessage());
         }
@@ -43,7 +60,38 @@ class CategoryController extends Controller
 
     public function detail(Request $req, $id) {
         try {
-            return $this->sendResponse(message: "Retrieve Category with ID::${id} success", data: Category::find($id));
+            return $this->sendResponse(message: "Retrieve Category with ID::$id success", data: Category::find($id));
+        } catch (\Throwable $th) {
+            return $this->sendError(message: $th->getMessage());
+        }
+    }
+
+    public function update(Request $req, $id) {
+        try {
+            $foundItem = Category::find($id);
+            if (!$foundItem) {
+                return $this->sendError(message: "Category with ID::$id not found!", statusCode: 404);
+            }
+
+            $checked = Gate::inspect('manage', $foundItem);
+            if (!$checked->allowed()) {
+                return $this->sendError(message: $checked->message(), statusCode: $checked->status());
+            }
+
+            $fields = $req->only(['name', 'image_url']);
+
+            $validated = Validator::make($fields, [
+                'name' => 'string|unique:categories,name|min:5',
+                'image_url' => 'string'
+            ]);
+
+            if ($validated->fails()) {
+                return $this->sendError(message: $validated->messages(), statusCode: 400);
+            }
+
+            $foundItem->update($fields);
+
+            return $this->sendResponse(message: "Update Category with id $id success.", statusCode: 201);
         } catch (\Throwable $th) {
             return $this->sendError(message: $th->getMessage());
         }
@@ -57,6 +105,11 @@ class CategoryController extends Controller
                 return $this->sendError(message: "Cannot find Category!", statusCode: 404);
             }
 
+            $checked = Gate::inspect('manage', $foundItem);
+            if (!$checked->allowed()) {
+                return $this->sendError(message: $checked->message(), statusCode: $checked->status());
+            }
+
             $foundItem->delete();
 
             return $this->sendResponse(message: "Remove Category with ID::${id} success");
@@ -64,6 +117,4 @@ class CategoryController extends Controller
             return $this->sendError(message: $th->getMessage());
         }
     }
-
-
 }
